pix_sig2pix~: make num-samples unsigned

+++ src/Pixes/pix_sig2pix~.h
@@ -53,7 +53,7 @@ protected:
   //////////
 
   // DSP perform
-  void perform(t_sample**signals, int count);
+  void perform(t_sample**signals, size_t count);
 
   //////////
   // Do the rendering
